fix(core): fix demo running error
